All test function move from vec2d.py to testdog.py
